requirements.txt
Django>=5.0
djangorestframework
markdown
django-filter
drf-yasg